+++ kahlan-config.php
@@ -2,3 +2,6 @@
 
 $args = $this->commandLine();
 $args->option('src', 'default', 'library/');
+
+unset($_SERVER['argc']);
+unset($_SERVER['argv']);